gig_bot/requirements.txt
python-telegram-bot==21.1.1
apscheduler==3.10.4
pytz==2024.1
Flask
gunicorn 
pytest
pytest-asyncio
